third_party/blink/web_tests/virtual/prefers-default-scrollbar-styles/fast/forms/textarea/textarea-resize-below-custom-resizer-size-expected.txt

This is a testharness.js-based test.
[FAIL] Test for resizing the TEXTAREA below the resizer size.
  assert_equals: expected 125 but got 15
Harness: the test ran to completion.
